sm6150-common: hidl: Drop wfdhdcp service
W libc    : Unable to set property "ctl.interface_start" to "android.hardware.drm@1.0::IDrmFactory/wfdhdcp": error code: 0x20

I hwservicemanager: Since android.hardware.drm@1.0::IDrmFactory/wfdhdcp is not registered, trying to start it as a lazy HAL.

I hwservicemanager: Tried to start android.hardware.drm@1.0::IDrmFactory/wfdhdcp as a lazy service, but was unable to. Usually this happens when a service is not installed, but if the service is intended to be used as a lazy service, then it may be configured incorrectly.